+++ CHANGELOG.md
@@ -7,3 +7,4 @@ CHANGELOG
 * Command system now works with `Update` object instead of `Message` (#14)
 * Drop support of php5
 * Drop support of symfony/symfony <= 3.4
+* Change bundle alias from `bo_shurik_telegram_bot` to `boshurik_telegram_bot`